GH-35337: [Go] ASAN tests fail with Go1.20+
