+++ go/go.mod
@@ -16,7 +16,7 @@
 
 module github.com/apache/arrow/go/v12
 
-go 1.18
+go 1.20
 
 require (
 	github.com/JohnCGriffin/overflow v0.0.0-20211019200055-46fa312c352c
